fix: edit barometer images
