+++ app/admin/components/dimensions.tsx
@@ -1,7 +1,7 @@
 import { Fieldset, TextInput, Group, ActionIcon, Tooltip, Stack } from '@mantine/core'
 import { IconTrash, IconSquareRoundedPlus } from '@tabler/icons-react'
 import { UseFormReturnType } from '@mantine/form'
-import { BarometerFormProps } from './add-card'
+import { type BarometerFormProps } from '@/app/types'
 
 interface DimensionsProps {
   form: UseFormReturnType<BarometerFormProps>
